+++ quickstart-connector/assembly.xml
