--- quickstart-connector/flinksql-connector-example/assembly.xml
@@ -1,39 +1,0 @@
-<assembly>
-    <id>bin</id>
-    <formats>
-        <format>tar.gz</format>
-    </formats>
-    <fileSets>
-        <fileSet>
-            <directory>assembly/bin</directory>
-            <outputDirectory>bin</outputDirectory>
-            <fileMode>0755</fileMode>
-        </fileSet>
-        <fileSet>
-            <directory>${project.build.directory}</directory>
-            <outputDirectory>lib</outputDirectory>
-            <fileMode>0755</fileMode>
-            <includes>
-                <include>*.jar</include>
-            </includes>
-            <excludes>
-                <exclude>original-*.jar</exclude>
-            </excludes>
-        </fileSet>
-        <fileSet>
-            <directory>assembly/conf</directory>
-            <outputDirectory>conf</outputDirectory>
-            <fileMode>0755</fileMode>
-        </fileSet>
-        <fileSet>
-            <directory>assembly/logs</directory>
-            <outputDirectory>logs</outputDirectory>
-            <fileMode>0755</fileMode>
-        </fileSet>
-        <fileSet>
-            <directory>assembly/temp</directory>
-            <outputDirectory>temp</outputDirectory>
-            <fileMode>0755</fileMode>
-        </fileSet>
-    </fileSets>
-</assembly>
