+++ quickstart-connector/src/main/java/com/streamxhub/streamx/test/flink/java/bean/Order.java
