+++ quickstart-connector/src/main/resources/logback.xml
