+++ quickstart-connector/src/main/scala/com/streamxhub/streamx/test/flink/scala/datastream/MongoSourceApp.scala
